requirements.txt
opencv-python
ultralytics

python-dotenv
langchain
langchainhub
langchain_openai